@@ -11,34 +11,34 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~8.1.3",
+    "@angular/animations": "~9.0.5",
     "@angular/cdk": "latest",
-    "@angular/common": "~8.1.3",
-    "@angular/compiler": "~8.1.3",
-    "@angular/core": "~8.1.3",
-    "@angular/forms": "~8.1.3",
+    "@angular/common": "~9.0.5",
+    "@angular/compiler": "~9.0.5",
+    "@angular/core": "~9.0.5",
+    "@angular/forms": "~9.0.5",
     "@angular/material": "latest",
     "@angular/material-moment-adapter": "latest",
-    "@angular/platform-browser": "~8.1.3",
-    "@angular/platform-browser-dynamic": "~8.1.3",
-    "@angular/router": "~8.1.3",
+    "@angular/platform-browser": "~9.0.5",
+    "@angular/platform-browser-dynamic": "~9.0.5",
+    "@angular/router": "~9.0.5",
     "moment": "^2.24.0",
     "ngx-daterangepicker-material": "^2.2.0",
     "ngx-socket-io": "^3.0.1",
-    "rxjs": "~6.4.0",
+    "rxjs": "~6.5.4",
     "saturn-datepicker": "^8.0.5",
-    "tslib": "^1.11.1",
-    "zone.js": "~0.9.1"
+    "tslib": "^1.10.0",
+    "zone.js": "~0.10.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.801.3",
-    "@angular/cli": "~8.1.3",
-    "@angular/compiler-cli": "~8.1.3",
-    "@angular/language-service": "~8.1.3",
+    "@angular-devkit/build-angular": "~0.900.5",
+    "@angular/cli": "~9.0.5",
+    "@angular/compiler-cli": "~9.0.5",
+    "@angular/language-service": "~9.0.5",
     "@types/jasmine": "~3.3.8",
     "@types/jasminewd2": "~2.0.3",
-    "@types/node": "~8.9.4",
-    "codelyzer": "^5.0.0",
+    "@types/node": "^12.11.1",
+    "codelyzer": "^5.1.2",
     "jasmine-core": "~3.4.0",
     "jasmine-spec-reporter": "~4.2.1",
     "karma": "~4.1.0",
@@ -49,6 +49,6 @@
     "protractor": "^5.4.3",
     "ts-node": "~7.0.0",
     "tslint": "~5.15.0",
-    "typescript": "~3.4.3"
+    "typescript": "~3.7.5"
   }
 }
